+++ src/Locator/UriLocatorInterface.php
@@ -9,12 +9,16 @@
  * file that was distributed with this source code.
  */
 
-namespace Webmozart\Puli\Repository;
+namespace Webmozart\Puli\Locator;
 
 /**
  * @since  1.0
  * @author Bernhard Schussek <bschussek@gmail.com>
  */
-class RemovalNotAllowedException extends \Exception
+interface UriLocatorInterface extends ResourceLocatorInterface
 {
+    /**
+     * @return string[]
+     */
+    public function getRegisteredSchemes();
 }